updated Player (client) to include mapID field; moved CoveySuperMapScene definition to CoveySuperMapScene.ts, created CoveySubMapScene in CoveySubMapScene.ts; added createMapScene function in WorldMap.tsx, updated WorldMap function to call createMapScene
